update:营养标准和原料部分的更新

+++ artemis-server/src/main/java/com/kuo/artemis/server/service/impl/FormulationServiceImpl.java
@@ -47,12 +47,15 @@ public class FormulationServiceImpl implements FormulationService {
     @Inject
     private FormulationMaterialMapper formulationMaterialMapper;
     /**
-     * 获取配方简略信息
+     * 获取配方基本信息
      * @param projectId
      * @return
      */
     public Response listFormulations(String projectId) {
         Response response = new Response();
+        if (projectId == null || "".equals(projectId)) {
+            return new Response(HttpStatus.BAD_REQUEST.value(), "参数不能为空");
+        }
 
         List<Formulation> formulationList = formulationMapper.selectByProjectId(Integer.valueOf(projectId));
 
@@ -74,12 +77,21 @@ public class FormulationServiceImpl implements FormulationService {
 
     public Response listFormulationNames(String projectId) {
 
-        List<String> formulationNameList = formulationMapper.selectFormulationNamesByProjectId(projectId);
+        if (projectId == null || "".equals(projectId)) {
+            return new Response(HttpStatus.BAD_REQUEST.value(), "参数不能为空");
+        }
+
+        List<Formulation> formulationNameList = formulationMapper.selectFormulationNamesByProjectId(Integer.valueOf(projectId));
 
         return new Response(formulationNameList, HttpStatus.OK.value(), "获取配方名称列表成功");
     }
 
     public Response getFormulationDetail(String formulationId) {
+
+        if (formulationId == null || "".equals(formulationId)) {
+            return new Response(HttpStatus.BAD_REQUEST.value(), "参数不能为空");
+        }
+
         FormulationResult result = new FormulationResult();
 
         Formulation formulation = formulationMapper.selectByPrimaryKey(Integer.valueOf(formulationId));
@@ -100,6 +112,10 @@ public class FormulationServiceImpl implements FormulationService {
     @Transactional(rollbackFor = Exception.class)
     public Response deleteFormulation(String formulationId) {
 
+        if (formulationId == null || "".equals(formulationId)) {
+            return new Response(HttpStatus.BAD_REQUEST.value(), "参数不能为空");
+        }
+
         Integer id = Integer.valueOf(formulationId);
 
         int resultMaterial = formulationMaterialMapper.deleteByFormulationId(id);
@@ -115,7 +131,7 @@ public class FormulationServiceImpl implements FormulationService {
     }
 
     /**
-     * TODO  ##BUG##  需保证原料id顺序为从小到大，原料营养指标的顺序符合Excel文件中 “营养指标设定” 那一页的顺序
+     * TODO  ##BUG##  需保证原料营养指标的顺序符合Excel文件中 “营养指标设定” 那一页的顺序
      * @param params
      * @return
      */
@@ -133,6 +149,11 @@ public class FormulationServiceImpl implements FormulationService {
         List<Material> materialList =  materialMapper.selectMaterialIndicators(materialIdList);
         NutritionStandard nutritionStandard = nutritionStandardMapper.selectByPrimaryKey(Integer.valueOf(nutritionStandardId));
 
+        if (materialIdList != null && nutritionStandard != null) {
+            //do nothing continue
+        } else {
+            return new Response(HttpStatus.BAD_REQUEST.value(),"原料或指标数据无效");
+        }
 
         //3.构建线性规划的参数
         //目标函数系数  即为各个原料的价格
@@ -178,24 +199,29 @@ public class FormulationServiceImpl implements FormulationService {
             //针对每一个营养标准指标构造约束方程
             Field field = fields[i];
             int materialSize = materialList.size();
-            List<Double> coefficientList = new ArrayList<Double>();
 
             if (field.isAnnotationPresent(NutritionIndicator.class)) {
 
-                for (int j = 0; j < materialSize; j++) {
-                    Material material = materialList.get(j);
-                    //Method method = material.getClass().getMethod("get" + field.getName(), BigDecimal.class);
-                    Map<String, Object> map = BeanUtil.beanToMap(material);
-                    Double coefficient = null;
-                    if (map.get(field.getName()) != null) {
-                        coefficient = ((BigDecimal) map.get(field.getName())).doubleValue();
-                        coefficientList.add(coefficient);
-                    } else {
-                        coefficientList.add(0.0);
+                NutritionIndicator annotation = field.getAnnotation(NutritionIndicator.class);
+                if (annotation.isIndicator() && BigDecimal.class.equals(field.getType())) {
+                    List<Double> coefficientList = new ArrayList<Double>();
+                    for (int j = 0; j < materialSize; j++) {
+                        Material material = materialList.get(j);
+                        //Method method = material.getClass().getMethod("get" + field.getName(), BigDecimal.class);
+                        Map<String, Object> map = BeanUtil.beanToMap(material);
+                        Double coefficient = null;
+                        if (map.get(field.getName()) != null) {
+                            coefficient = ((BigDecimal) map.get(field.getName())).doubleValue();
+                            coefficientList.add(coefficient);
+                        } else {
+                            coefficientList.add(0.0);
+                        }
                     }
+
+                    constraintLists.add(coefficientList);
                 }
 
-                constraintLists.add(coefficientList);
+
             }
         }
     }
@@ -203,6 +229,8 @@ public class FormulationServiceImpl implements FormulationService {
     private List<FormulationMaterial> getFormulationMaterials(List<Material> materialList, LinearProgrammingResult result) {
         List<FormulationMaterial> formulationMaterialList = new ArrayList<FormulationMaterial>();
         List<Double> resultValue = result.getVariableValueList();
+
+        //TODO decimalFormat改为单例模式
         DecimalFormat decimalFormat = new DecimalFormat("0.000000");
         for (int i = 0; i < resultValue.size(); i++) {
             FormulationMaterial formulationMaterial = new FormulationMaterial();
@@ -227,12 +255,15 @@ public class FormulationServiceImpl implements FormulationService {
         for (int i = 0; i < fields.length; i++) {
             Field field = fields[i];
             if (field.isAnnotationPresent(NutritionIndicator.class)) {
-                nutritionNameList.add(field.getAnnotation(NutritionIndicator.class).name());
-                BigDecimal value = (BigDecimal) map.get(field.getName());
-                if (value != null) {
-                    nutritionContentIdealList.add(value.doubleValue());
-                } else {
-                    nutritionContentIdealList.add(-1D);
+                NutritionIndicator annotation = field.getAnnotation(NutritionIndicator.class);
+                if (annotation.isIndicator() && BigDecimal.class.equals(field.getType())) {
+                    nutritionNameList.add(field.getAnnotation(NutritionIndicator.class).name());
+                    BigDecimal value = (BigDecimal) map.get(field.getName());
+                    if (value != null) {
+                        nutritionContentIdealList.add(value.doubleValue());
+                    } else {
+                        nutritionContentIdealList.add(-1D);
+                    }
                 }
 
             }
