+++ artemis-server/src/main/java/com/kuo/artemis/server/service/impl/MaterialServiceImpl.java
@@ -1,6 +1,7 @@
 package com.kuo.artemis.server.service.impl;
 
 import com.kuo.artemis.server.core.dto.Response;
+import com.kuo.artemis.server.core.helper.DataHelper;
 import com.kuo.artemis.server.dao.MaterialMapper;
 import com.kuo.artemis.server.entity.Material;
 import com.kuo.artemis.server.service.MaterialService;
@@ -22,24 +23,40 @@ public class MaterialServiceImpl implements MaterialService {
     private MaterialMapper materialMapper;
 
     public Response listMaterials(String userId) {
+        if (userId == null || "".equals(userId)) {
+            return new Response(HttpStatus.BAD_REQUEST.value(), "参数不能为空");
+        }
+
         List<Material> materialList = materialMapper.selectByUserId(Integer.valueOf(userId));
+        //List<List<String>> data = DataHelper.objectListToArrayList(materialList);
 
         return new Response(materialList, HttpStatus.OK.value(), "获取原料列表成功");
     }
 
     public Response listMaterialsBrief(String userId) {
+        if (userId == null || "".equals(userId)) {
+            return new Response(HttpStatus.BAD_REQUEST.value(), "参数不能为空");
+        }
+
         List<Material> materialList = materialMapper.selectBriefByUserId(Integer.valueOf(userId));
 
         return new Response(materialList, HttpStatus.OK.value(), "获取原料简略列表成功");
     }
 
     public Response getMaterialDetail(String materialId) {
+        if (materialId == null || "".equals(materialId)) {
+            return new Response(HttpStatus.BAD_REQUEST.value(), "参数不能为空");
+        }
+
         Material material = materialMapper.selectByPrimaryKey(Integer.valueOf(materialId));
 
         return new Response(material, HttpStatus.OK.value(), "获取原料成功");
     }
 
     public Response updateMaterialDetail(Material material, String userId) {
+        if (userId == null || "".equals(userId)) {
+            return new Response(HttpStatus.BAD_REQUEST.value(), "参数不能为空");
+        }
 
         material.setUserId(Integer.valueOf(userId));
         int result = materialMapper.updateByPrimaryKeySelective(material);
@@ -53,6 +70,10 @@ public class MaterialServiceImpl implements MaterialService {
     }
 
     public Response deleteMaterial(String materialId) {
+        if (materialId == null || "".equals(materialId)) {
+            return new Response(HttpStatus.BAD_REQUEST.value(), "参数不能为空");
+        }
+
         int result = materialMapper.deleteByPrimaryKey(Integer.valueOf(materialId));
 
         if (result > 0) {
@@ -63,6 +84,9 @@ public class MaterialServiceImpl implements MaterialService {
     }
 
     public Response createNewMaterial(Material material, String userId) {
+        if (userId == null || "".equals(userId)) {
+            return new Response(HttpStatus.BAD_REQUEST.value(), "参数不能为空");
+        }
 
         material.setUserId(Integer.valueOf(userId));
         int result = materialMapper.insert(material);
